# contracts/openapi/api.yaml
openapi: 3.1.0
info:
  title: Dorahub Internal API
  version: 0.18.0
  description: |
    JSON uses lowerCamelCase property names and RFC 3339 UTC timestamps.
    Collection endpoints use zero-based pages and the shared page, size, sort and filter parameters.
servers:
  - url: /
paths:
...
  /api/v1/ratings:
    get:
      operationId: getLadder
      summary: Ladder of one format, strongest first
      parameters:
        - name: format
          in: query
          required: true
          description: Ladder key, for example "hanchan". Formats are rated separately.
          schema: { type: string, minLength: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Ladder rows ordered by rating
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LadderEntryView"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/ratings/me/summary:
    get:
      operationId: getMyRatingSummary
      summary: Own games, places and rating per format
      parameters:
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: One row per format the player has finished games in
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormatSummaryView"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/ratings/me:
    get:
      operationId: getMyRatingChanges
      summary: Own rating changes, newest first
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Rating changes of the signed-in player
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingChangeView"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DORAHUB_SESSION
  headers:
    CorrelationId:
      description: Request correlation identifier. The server returns a safe supplied value or generates one.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,64}$"
  parameters:
...
    RequestCorrelationId:
      name: X-Correlation-Id
      in: header
      description: Optional caller-generated request correlation identifier.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,64}$"
    ClientSource:
      name: X-Client-Source
      in: header
      description: Low-cardinality client channel, without device or user identifiers.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,32}$"
...
  responses:
    BadRequest:
      $ref: "#/components/responses/ProblemResponse"
      description: Malformed request (`request.invalid`).
    Unauthorized:
      $ref: "#/components/responses/ProblemResponse"
      description: Authentication is required (`authentication.required`).
...
    ProblemResponse:
      description: Request failed.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    LadderEntryView:
      type: object
      required: [accountId, rating, games]
      properties:
        accountId: { type: string, format: uuid }
        nickname:
          type: [string, "null"]
          description: Absent for an account that no longer exists.
        rating: { type: integer }
        games: { type: integer, minimum: 1 }
    FormatSummaryView:
      type: object
      required: [format, rating, games, places, averagePlace]
      properties:
        format: { type: string }
        rating: { type: integer }
        games:
          type: integer
          minimum: 1
          description: Games that still count; a game taken off the ladder is not among them.
        places:
          type: array
          description: How many times the player finished first, second, third and fourth.
          items: { type: integer, minimum: 0 }
          minItems: 4
          maxItems: 4
        averagePlace: { type: number }
...
    RatingChangeView:
      type: object
      required: [tableId, format, place, delta, ratingAfter, at]
      properties:
        tableId: { type: string, format: uuid }
        format: { type: string }
        place: { type: integer, minimum: 1, maximum: 4 }
        delta:
          type: integer
          description: Change of rating for this game. Deltas of one game sum to zero.
        ratingAfter: { type: integer }
        at: { type: string, format: date-time }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string, pattern: "^[a-z][a-z0-9_]*(\\.[a-z][a-z0-9_]*)+$" }
        correlationId:
          type: string
          pattern: "^[A-Za-z0-9._-]{1,64}$"
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"